move theme into own package and revert some formatting's
